audio: build with GC/1.3 and clear the load flags plainly
Under GC/2.0 a 32-bit single-bit clear always folds to rlwinm, so the
pending-flag clears carried a ~(u64) cast to force retail's li/and pair.
That cast materialises a zero-extension high word, and in
streamsLoadedCallback it coalesces with the loop's own zero and survives
eight instructions too early.

GC/1.3 emits retail's li/and shape from the plain 32-bit clear, so the casts
come out and streamsLoadedCallback matches: 97.05714 -> 100.0 (280 bytes).

The cost is audioSetVolumes (108 bytes), which calls sndMasterVolume through
a cast function pointer to dodge the musyx prototype; GC/2.0 sees through
that and emits bl, GC/1.3 always goes through the counter register. Keeping
a direct call under 1.3 would need a conflicting prototype in scope, so the
unit is one function short either way -- this way the matched one is larger.

Unit 99.91717 -> 99.85525, 37/38 functions, net +172 bytes of matched code.
All data sections unchanged.

+++ src/main/audio.c
@@ -76,7 +76,7 @@ int gAudioArqRequestIndex;
 
 /* gAudioPendingLoadFlags / gAudioCompletedLoadFlags: one bit per async
  * resource load, set when enqueued and cleared/mirrored when the load
- * callback fires. (Pending clears use a 64-bit mask: ~(u64)FLAG.) */
+ * callback fires. */
 #define AUDIO_LOAD_MUSIC_TRIGGERS 0x1   /* music trigger table */
 #define AUDIO_LOAD_SFX_TRIGGERS   0x2   /* sfx trigger table */
 #define AUDIO_LOAD_STREAMS        0x4   /* stream table */
@@ -293,7 +293,7 @@ void sampleBufferSLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_S_SAMPLE_BUF;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_S_SAMPLE_BUF;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_S_SAMPLE_BUF;
     }
 }
@@ -315,7 +315,7 @@ void sampleDirectorySLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_S_SAMPLE_DIR;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_S_SAMPLE_DIR;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_S_SAMPLE_DIR;
     }
 }
@@ -337,7 +337,7 @@ void projectDataSLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_S_PROJECT;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_S_PROJECT;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_S_PROJECT;
     }
 }
@@ -359,7 +359,7 @@ void poolDataSLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_S_POOL;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_S_POOL;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_S_POOL;
     }
 }
@@ -381,7 +381,7 @@ void sampleBufferMLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_M_SAMPLE_BUF;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_M_SAMPLE_BUF;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_M_SAMPLE_BUF;
     }
 }
@@ -403,7 +403,7 @@ void sampleDirectoryMLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_M_SAMPLE_DIR;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_M_SAMPLE_DIR;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_M_SAMPLE_DIR;
     }
 }
@@ -425,7 +425,7 @@ void projectDataMLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_M_PROJECT;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_M_PROJECT;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_M_PROJECT;
     }
 }
@@ -448,7 +448,7 @@ void poolDataMLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_M_POOL;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_M_POOL;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_M_POOL;
     }
 }
@@ -473,7 +473,7 @@ void streamsLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_STREAMS;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_STREAMS;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_STREAMS;
         stream = gStreamsData;
         streamCount = gStreamsCount;
@@ -503,7 +503,7 @@ void sfxTriggersLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_SFX_TRIGGERS;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_SFX_TRIGGERS;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_SFX_TRIGGERS;
     }
 }
@@ -525,7 +525,7 @@ void musicTriggersLoadedCallback(s32 status, DVDFileInfo* fileInfo)
         saved = mmSetFreeDelay(0);
         mm_free(fileInfo);
         mmSetFreeDelay(saved);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_MUSIC_TRIGGERS;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_MUSIC_TRIGGERS;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_MUSIC_TRIGGERS;
     }
 }
@@ -847,7 +847,7 @@ static void MIDIWADLoadedCallback(s32 status, DVDFileInfo* fileInfo) {
     } else {
         DVDClose(fileInfo);
         mm_free(fileInfo);
-        gAudioPendingLoadFlags &= ~(u64)AUDIO_LOAD_MIDI_WAD;
+        gAudioPendingLoadFlags &= ~AUDIO_LOAD_MIDI_WAD;
         gAudioCompletedLoadFlags |= AUDIO_LOAD_MIDI_WAD;
     }
 }
